Version_03/requirements.txt
streamlit
rich
groq
python-dotenv